requirements.txt
Pillow
pyautogui
jamo